# docs/protocol/local-admin-api-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Private Sync Gate Local Administration API",
    "version": "1.0.0",
    "description": "Loopback-only, owner-controlled administration API for the local daemon."
  },
  "servers": [
    { "url": "http://127.0.0.1:47820", "description": "Local daemon loopback listener" }
  ],
  "paths": {
...
    "/api/v1/browser-session/bootstrap": {
      "post": {
        "operationId": "exchangeBrowserSession",
        "summary": "Exchange a one-use token for a same-origin browser session",
        "security": [],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BrowserSessionBootstrap" } } } },
        "responses": { "200": { "description": "Authenticated sanitized browser session document", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/BrowserSessionDocument" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "503": { "$ref": "#/components/responses/Unavailable" } }
      }
    },
...
    "/api/v1/status": {
      "get": {
        "operationId": "getStatus",
        "summary": "Read daemon status",
        "security": [{ "bearerAuth": [] }],
        "responses": { "200": { "description": "Current status", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/StatusResponse" } } } }, "401": { "$ref": "#/components/responses/Unauthorized" }, "503": { "$ref": "#/components/responses/Unavailable" } }
      }
    },
...
    "/api/v1/federation/nodes": {
      "get": {
        "operationId": "listFederatedNodes",
        "summary": "List sanitized signed status replicas for paired nodes with active read-only grants",
        "security": [{ "bearerAuth": [] }, { "browserSession": [] }],
        "responses": { "200": { "description": "Visible paired-node status replicas", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/FederatedNodePage" } } } }, "401": { "$ref": "#/components/responses/Unauthorized" }, "503": { "$ref": "#/components/responses/Unavailable" } }
      }
    },
...
    "/api/v1/projects": {
      "get": {
        "operationId": "listProjects",
        "summary": "List registered Agent Projects without local root paths",
        "security": [{ "bearerAuth": [] }],
        "parameters": [{ "$ref": "#/components/parameters/Limit" }, { "$ref": "#/components/parameters/Cursor" }],
        "responses": { "200": { "description": "Project page", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ProjectPage" } } } }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "503": { "$ref": "#/components/responses/Unavailable" } }
      }
    },
...
    "/api/v1/orchestration/workers": { "get": { "operationId": "listWorkers", "summary": "List versioned worker inventory", "security": [{ "bearerAuth": [] }, { "browserSession": [] }], "parameters": [{ "$ref": "#/components/parameters/Limit" }, { "$ref": "#/components/parameters/Cursor" }], "responses": { "200": { "description": "Worker page" }, "400": { "$ref": "#/components/responses/BadRequest" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "503": { "$ref": "#/components/responses/Unavailable" } } } },
...
    "/api/v1/orchestration/projects": { "get": { "operationId": "listLocalProjects", "summary": "List sanitized local project operation status", "security": [{ "bearerAuth": [] }, { "browserSession": [] }], "parameters": [{ "$ref": "#/components/parameters/Limit" }, { "$ref": "#/components/parameters/Cursor" }], "responses": { "200": { "description": "Bounded local project page", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/LocalProjectPage" } } } }, "401": { "$ref": "#/components/responses/Unauthorized" }, "503": { "$ref": "#/components/responses/Unavailable" } } } },
...
    "/api/v1/orchestration/nodes": { "get": { "operationId": "listOrchestrationNodes", "summary": "List sanitized node inventory", "security": [{ "bearerAuth": [] }, { "browserSession": [] }], "parameters": [{ "$ref": "#/components/parameters/Limit" }, { "$ref": "#/components/parameters/Cursor" }], "responses": { "200": { "description": "Node page" }, "401": { "$ref": "#/components/responses/Unauthorized" }, "503": { "$ref": "#/components/responses/Unavailable" } } } },
...
    "/api/v1/projects/{project_id}/assignments/{assignment_id}/controls": { "post": { "operationId": "controlAssignment", "summary": "Pause, resume, cancel, fail, retry, evaluate, or reassign idempotently", "security": [{ "bearerAuth": [] }, { "browserSession": [] }], "parameters": [{ "$ref": "#/components/parameters/ProjectID" }, { "$ref": "#/components/parameters/AssignmentID" }], "responses": { "200": { "description": "Assignment detail" }, "400": { "$ref": "#/components/responses/BadRequest" }, "409": { "$ref": "#/components/responses/Conflict" }, "503": { "$ref": "#/components/responses/Unavailable" } } } },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "local-admin-token" },
      "browserSession": { "type": "apiKey", "in": "cookie", "name": "syncgate_browser_session" }
    },
    "parameters": {
      "Limit": { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 200, "default": 50 } },
      "Cursor": { "name": "cursor", "in": "query", "schema": { "type": "string", "maxLength": 512 } },
...
      "ProjectID": { "name": "project_id", "in": "path", "required": true, "schema": { "type": "string", "minLength": 1, "maxLength": 256 } },
...
      "AssignmentID": { "name": "assignment_id", "in": "path", "required": true, "schema": { "type": "string", "minLength": 1, "maxLength": 256 } },
...
    },
    "responses": {
      "BadRequest": { "description": "Invalid request", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Unauthorized": { "description": "Missing or invalid local credential", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
...
      "Conflict": { "description": "Command conflicts with current state", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } },
      "Unavailable": { "description": "Daemon service unavailable", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } } }
    },
    "schemas": {
      "Error": { "type": "object", "required": ["error"], "properties": { "error": { "type": "object", "required": ["code", "message"], "properties": { "code": { "type": "string" }, "message": { "type": "string", "maxLength": 512 }, "request_id": { "type": "string", "maxLength": 128 } } } } },
...
      "BrowserSessionBootstrap": { "type": "object", "additionalProperties": false, "required": ["bootstrap_token"], "properties": { "bootstrap_token": { "type": "string", "minLength": 32, "maxLength": 128 } } },
      "BrowserSessionDocument": { "type": "object", "additionalProperties": false, "required": ["api_version", "health", "status", "capabilities", "csrf_token", "session_expires_at"], "properties": { "api_version": { "type": "string", "const": "v1" }, "health": { "type": "string", "enum": ["ok", "unavailable"] }, "status": { "$ref": "#/components/schemas/StatusResponse" }, "capabilities": { "type": "array", "maxItems": 9, "items": { "type": "string", "enum": ["node.health.read", "node.status.read", "project.visibility.read", "task.readiness.read", "assignment.visibility.read", "result-budget-incident.read", "paired-node-status.read", "worker-node.inventory.read", "operator.controls.write"] } }, "csrf_token": { "type": "string", "minLength": 32, "maxLength": 128 }, "session_expires_at": { "type": "string", "format": "date-time" } } },
      "StatusResponse": { "type": "object", "required": ["status", "api_version", "device_id", "fingerprint", "started_at", "lifecycle", "active_share_count", "peer_executor_enabled", "queue"], "properties": { "status": { "type": "string", "enum": ["starting", "running", "draining"] }, "api_version": { "type": "string", "const": "v1" }, "device_id": { "type": "string" }, "fingerprint": { "type": "string" }, "started_at": { "type": "string", "format": "date-time" }, "lifecycle": { "type": "string", "enum": ["starting", "running", "draining"] }, "active_share_count": { "type": "integer", "minimum": 0 }, "peer_executor_enabled": { "type": "boolean" }, "queue": { "$ref": "#/components/schemas/QueueSummary" } } },
      "QueueSummary": { "type": "object", "required": ["pending", "running", "paused", "failed"], "properties": { "pending": { "type": "integer", "minimum": 0 }, "running": { "type": "integer", "minimum": 0 }, "paused": { "type": "integer", "minimum": 0 }, "failed": { "type": "integer", "minimum": 0 } } },
...
      "PageInfo": { "type": "object", "required": ["limit", "has_more"], "properties": { "limit": { "type": "integer", "minimum": 1, "maximum": 200 }, "next_cursor": { "type": "string", "maxLength": 512 }, "has_more": { "type": "boolean" } } },
...
      "FederatedStatusCount": { "type": "object", "additionalProperties": false, "required": ["state", "count"], "properties": { "state": { "type": "string", "maxLength": 64 }, "count": { "type": "integer", "minimum": 0 } } },
      "FederatedProjectSummary": { "type": "object", "additionalProperties": false, "required": ["project_id", "display_name", "scheduler_state"], "properties": { "project_id": { "type": "string", "maxLength": 256 }, "display_name": { "type": "string", "maxLength": 128 }, "scheduler_state": { "type": "string", "enum": ["running", "paused", "disabled", "draining", "not_selected", "unknown"] }, "assignment_counts": { "type": "array", "maxItems": 16, "items": { "$ref": "#/components/schemas/FederatedStatusCount" } }, "gate_counts": { "type": "array", "maxItems": 16, "items": { "$ref": "#/components/schemas/FederatedStatusCount" } }, "accepted_history_watermark": { "type": "string", "pattern": "^[a-f0-9]{64}$" } } },
      "FederatedNode": { "type": "object", "additionalProperties": false, "required": ["device_id", "display_name", "protocol", "revision", "watermark", "health", "lifecycle", "connectivity", "observed_at", "expires_at", "grant_expires_at", "projects"], "properties": { "device_id": { "type": "string", "maxLength": 256 }, "display_name": { "type": "string", "maxLength": 256 }, "protocol": { "type": "string", "const": "syncgate-node-status-v1" }, "revision": { "type": "integer", "minimum": 1 }, "watermark": { "type": "string", "pattern": "^[a-f0-9]{64}$" }, "health": { "type": "string", "enum": ["healthy", "degraded", "unhealthy", "unknown"] }, "lifecycle": { "type": "string", "enum": ["starting", "running", "draining", "stopped", "error"] }, "connectivity": { "type": "string", "enum": ["online", "offline"] }, "observed_at": { "type": "string", "format": "date-time" }, "expires_at": { "type": "string", "format": "date-time" }, "grant_expires_at": { "type": "string", "format": "date-time" }, "projects": { "type": "array", "maxItems": 200, "items": { "$ref": "#/components/schemas/FederatedProjectSummary" } } } },
      "FederatedNodePage": { "type": "object", "additionalProperties": false, "required": ["items"], "properties": { "items": { "type": "array", "maxItems": 200, "items": { "$ref": "#/components/schemas/FederatedNode" } } } },
...
      "ProjectProjectionState": { "type": "object", "additionalProperties": false, "required": ["history", "insights"], "properties": { "history": { "type": "string", "enum": ["available", "unavailable", "rebuilding"] }, "insights": { "type": "string", "enum": ["available", "unavailable", "rebuilding", "stale"] }, "checkpoint_at": { "type": "string", "format": "date-time" } } },
      "Project": { "type": "object", "additionalProperties": false, "required": ["id", "share_id", "name", "authority_device_id", "registered_at", "projection"], "properties": { "id": { "type": "string" }, "share_id": { "type": "string" }, "name": { "type": "string" }, "authority_device_id": { "type": "string" }, "registered_at": { "type": "string", "format": "date-time" }, "projection": { "$ref": "#/components/schemas/ProjectProjectionState" } } },
      "ProjectPage": { "type": "object", "required": ["items", "page"], "properties": { "items": { "type": "array", "maxItems": 200, "items": { "$ref": "#/components/schemas/Project" } }, "page": { "$ref": "#/components/schemas/PageInfo" } } },
...
      "StatusCount": { "type": "object", "additionalProperties": false, "required": ["state", "count"], "properties": { "state": { "type": "string", "maxLength": 64 }, "count": { "type": "integer", "minimum": 0 } } },
      "LocalProject": { "type": "object", "additionalProperties": false, "required": ["project_id", "display_name", "registered_at", "selected", "execution_authorized", "scheduling_enabled", "max_concurrent", "scheduler_state", "assignment_counts", "gate_counts"], "properties": { "project_id": { "type": "string", "maxLength": 256 }, "display_name": { "type": "string", "maxLength": 256 }, "registered_at": { "type": "string", "format": "date-time" }, "selected": { "type": "boolean" }, "execution_authorized": { "type": "boolean" }, "scheduling_enabled": { "type": "boolean" }, "max_concurrent": { "type": "integer", "minimum": 1, "maximum": 2 }, "scheduler_state": { "type": "string", "enum": ["not_selected", "disabled", "paused", "running", "draining"] }, "assignment_counts": { "type": "array", "maxItems": 12, "items": { "$ref": "#/components/schemas/StatusCount" } }, "gate_counts": { "type": "array", "maxItems": 4, "items": { "$ref": "#/components/schemas/StatusCount" } } } },
      "LocalProjectPage": { "type": "object", "additionalProperties": false, "required": ["items", "page"], "properties": { "items": { "type": "array", "maxItems": 200, "items": { "$ref": "#/components/schemas/LocalProject" } }, "page": { "$ref": "#/components/schemas/PageInfo" } } },
...
    }
  }
}